move claim_vault method to its own file
